Fix exporting Markdown formatted content
There was an inconsistent number of white spaces at
the end of metadata lines in a Markdown formatted
output.

Markdown format requires two spaces if a newline is
requested. This is used for example with the metadata
section. If the metadata field was empty, it produced
three spaces. Also the last field `version` had only
one space at the end of the line.

Every line in the Markdown formatted data in the Meta
section now has consistent number of two white spaces
at the end of line. This applies if the meta field is
empty or if it has any content.

This did not cause problems at least in online Mkdn
formatters because three spaces was working. This a
fix to produce consistent output for example to
post process the produced files.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>