Add code commit to clarify Markdown template
When a new Markdown formatted snippet is created, there
is a prefilled data set for a template to help to fill
the content. The brief field does not have a dot at the
end of sentence. This is because the brief field also
contains the group field and adding a dot would not
look nice.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>