Add document for required setuptools version
Snippy setup.py uses environment markers for example for
Pylint. There are comments in Internet that these
would not work because wheel would strip these out in
install_requires. These are old comments.

This is fixed in setuptools 36.2.1 [1]. This is noted
incorrectly in previous commit message which contains
a note about the release of setuptools that introduced
the environment markers. But the fix to support these
is in version 36.2.1.

This is already taken into account in Snippy pyproject
toml file. This is more of an documentation about the
needed versions for setup related tools for the Snippy
project.

[1] https://github.com/pypa/setuptools/issues/1087

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>