Fix first REST API server response
If there was a failure that created a Cause during the
server startup, the first response from the server
contained error response even when the request was
successfully executed.

The cause was not reset before the server was started
and the global cause codes were read in the first
request. This caused incorrect response.

This is more relevant only when the latest addition
of reading default content before the server starts.
The case may be that the content is already imported
for example in PostgreSQL database. This would create
situation that adds a cause code.

Previously there may have been some cases that would
have generated the same situation. How ever, these
cases are not know.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>