Fix (partially) unnecessary return of collection
Import and create operations use the same storage API to
insert content. The create operation requires that the
inserted content is returned. But in case of import, there
is no use case for this.

There is now unnecessary variable in import_content method.
Previously this was returned from the storage but that is
now removed to clear the Storage() users. The problem of
unnecessary return in storage still remains for the import
operation.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>