Add support for environment variable configuration
It is now possible to use environment variables to configure
server and storage options. Only these two option groups
support environment variables.

The configuration precedence is:

  1. command line options
  2. environment variables
  3. hard coded defaults

The intention is to support container and bare metal based
server and possible usage of other than Sqlite as a storage
backend through environment variables. It is a common pattern
to define configuration through environment variables with
Docker containers.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>