Add better logs for argument parsing failures
If the tool is run with -vv or --debug options, the argument
parsing failure is now printed at the end of the logs. This
does not affect normal run without -vv or --debug options.

This required overriding ArgumentParser error handling.

Overkill? Bad design to override ArgumentParser? Minimal
"improvements"?

Hell Yeah! But this was on the list. Also learned how to get
the ArgumentParser error message and passing **kwargs and
trying to modify dictionary through **kwargs (not working)

This may be a bit better than [1] because this allows reading
the error message and doing what ever with it with minimal
changes to ArgumentParser.

[1] https://stackoverflow.com/a/5943331

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>