Add environment variable support for debug options
Environment variables can now be used for support options
on top of server and storage options. The support options
allow setting debug options like JSON logs. This is useful
if for example container based server crashes and debug
options are required to remain over reboot.

Fixed parsing of environment variables if the value was
integer. Previously this was not a problem because the
server and storage options did not have integer values.
Environment variables are strings and they must be
converted to integers if the configuration value is an
integer value.

Refactored the Cli and argument parser usage. Arguments
with hyphens are automatically converted to underscores
by the ArgumentParser. There is no need to use the `dest`
option. This reduced bloat and line lengths.

NOTE! All internal variables in Cli that are using the
      environment variables, must have same variable
      as the command line option. Otherwise environment
      variables are not read correctly.

      This is not a problem because the Cli options are
      read to another internal structure that allows
      longer names for example for -q or -vv options.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>