Add JSON logs by default in Docker container
The container now sets JSON logs by default. The logs can be
still controlled with -vv, --debug or -q options when the
server is started. But on top of these command line options,
there is no need to define JSON logs to be used in container.

The JSON logging is common pattern for container based micro
services which write logs to stdout where they get forwarded
to sink which usually supports JSON formatting to be able to
post process the logs nicely.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>