Fix duplicated content field values
If user created or updated content fields with duplicated
values, the duplicated values were stored. It is assumed
that user does not want to store same values twice. There
are no known use case to store same tag, group or link twice.

This fix stores only unique values to all content fields
which are stored as a list.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>