Edit test case comments
Some of the test cases comments were improved and harmonized.
There is now also few rules how to write command line options
for code and test comments and for user guides and documents.

The intention is to harmonize the test case documentation
in order to auto generate test documents. For example a test
case header should allow reader to gain enough information
what is tested in order to find relevant test cases.

It is expected that the Markdown or HTTP formats are used
and thus for example Markdown code span is defined to be
used with the command line options.

See the details how to write command line options in various
documents from the SCRIBBLING change document (that is moved
to proper document someday).

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>