Fix tool pre-release version to follow PEP440
Changed the pre-release version to follow syntax defined
in PEP440. This prevents errors and warnings from setup
tools.

The version semantic for pre-release follow syntax x.yaN.
This allows same length version number than final release.
The test cases check REST API response length and error
cases contain the tool version number. This helps to keep
the REST API response same length and they do not have
to be updated when release changes 0.10a0 to 0.10.0.

Command example that produce failures with version numbers
that do not follow PEP440 are below. If commands produce
errors, it causes problems for example with Makefile that
will rely on the output of these commands in the future.

  $ python setup.py --version
  $ python setup.py --name

[1] https://www.python.org/dev/peps/pep-0440/#pre-releases

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>