Edit document how to install PyPy dependencies
It seems that the PyPy for Python 2.7 does not support
implementation_name=="PyPy" [1]. This would mean that
it is not possible to compine extras_server to other
than Python3 based PyPy. It seems that with PyPy for
Python 2, any value in implementation_name tries to
install it dependency.

Added also experiments to try to get common $(PYPY)
to Makefile but it turned out to be difficult.

The solution how to do this is not clear. This commit
just adds the current investigation.

Note that for example upgrading setuptools and/or
pip with PyPy 5.5 (Python 3.3) breaks because there
are modules in latest pip that require at least
Python 3.4.

[1] https://www.python.org/dev/peps/pep-0421/#pypy

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>