Refactor Makefile to improve PyPy installation (WIP)
The PyPy and psycopg2cffi installations should be seamless
to user. The problem is that the psycopg2 and psycopgcffi
installations have dependencies which are not installed
with pip. This causes problems if the dependencies are
not installed.

There is now initial implementation to check if required
packages are installed on DNF based package manager. This
does not yet quite work since the test status is not yet
checked to terminate the make target with proper error
if the dependencies are missing.

There is still target to try to combine the setup.py
server targets so that the psycopg2* would be properly
installed with Python and PyPy targets.

Since Alpine does not support PyPy (or PyPy does not
support Alpine) it is not likely that the PyPY server
can be installed on this base image.

Makefile stuff turned out to be harder than expected so
the current implementation may not be perfect.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>