Fix Makefile backslash alignment for 8 space tabs
The Makefile function with backslashes did not look nice when
it was displayed with 8 space tabs. Now the alignment is done
for 8 space which seems to be Linux default for Makefile.

It is not know how to do this so that it is correct with any
tab width.

Removed unnecessary global parameter from the Makefile.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>