Fix updating of Markdown formatted snippet
If a snippet had multiple commands but only few of them had
comment after the command, the Markdown formatted update in
editor was not consistent. The editor displayed the commands
properly with or without the comments translated to Markdown
headers. But for user this was considered strange since some
of the commands were aligned differently.

This did not produce parsing error. Just the presentation
format of the data was considered bad.

There is now temporary and automatically added header/comment
for snippet commands that do not have comments. This temporary
header is not stored into database. It is just displayed for
user when content is updated with editor. The same temporary
comment/header is added for exported data. But when exported
data with temporary comment is imported, comment is removed.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>