Fix test case comments
Highlighting in comments must be made with two backticks.
The reason is that the rst format requires two backticks
to highlight words and phrases. A single backtick should
be implicit reference [1].

The code documentation is generated in rst format and
not in Markdown format which has a single backtick for
code block quote.

[1] https://github.com/coala/projects/issues/391

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>