Add gzip compression support for REST API responses
It is now possible to send HTTP requests with Accept-Encoding
header which contains value `gzip`. This causes the REST API
server to respond with a compressed body and header which
has the Content-Encoding set to `gzip` if compression was
done.

Client must always check if the Content-Encoding header is
set and if the value is `gzip`. Errors and the hello API
endpoints are never compressed. It is assumed that errors
and the hello API responses are required always in text
format to help troubleshooting.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>