Fix creating new content with prefilled templates
If content is created with an editor, the editor template has
a prefilled example content. If the example content was not
removed, it was stored with the content if the mandatory field
was filled correctly.

Now the example content presented by editor by default is
removed automatically from the text template from text and
Markdown formatted data. This allows user just to fill the
required fields and leave rest of the fields with examples
that are now automatically removed when the content is saved.

Previously text or Markdown template was identified only
when the template was not changed at all.

This change also simplified empty template digest management.
In case of snippets and references, the digest of an empty
template is now the same with a text and Markdown templates.
In case of a solution content there is still a difference
because text and Markdown formatted solution templates have
different data content.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>