Add support for name, source and versions fields
It is now possible to create and update content from text
and Markdown formatted files from command line with the
new fields.

The support for these fields was already added into the REST
API server. The missing part was parsing these fields from
command line input.

The fields are not available from command line interface.
Updated must be made with editor to give a value for these
fields.

The `name` field is intended to store a single name for
the content. For example bookmark link set can contain
short and unique name to identify the link in a bookmark.

The `source` field is intended to store the source where
the content was read. For example in future there may be
a functionality that imports snippets like tldr-pages. In
this kind of cases this field would contain a link to the
reference where the data was read and imported.

The `versions` field is intended to store version data for
the content. In most cases the content examples are specific
to versions like Python 3 and Python 3. This field is to
store version dependencies for the given content.

The `versions` field must have key value pairs with supported
operator. The operators are =, <, >, <=, >= and !=.

Currently the REST API returns 200 and stores the content
with valid versions if there is a misuse for the rules. This
is going to change in future.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>