Fix default value setting to groups field
If CLI or API source did not set value for the groups fields,
it was left empty. The intention has been that there is always
at least the `default` value in groups. There is not actual
problem in the code if the field is left empty. Just that the
design target has been enforcing at least some value in this
field.

The original thought why this was the case might have had
something to do with searching and content presentation in
terminal. Event though the implementation enforces the default,
there would not be real objections currently to leave the field
empty.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>