Fix missing pyproject.toml build-backend
The PEP-518 mandated that all packages having the pyproject.toml
must specify the build-backend section [1]. This field was not
defined in Snippy pyproject.toml file previously. There has
been updates at least to setuptools [2] to allow compatibility
mode without this parameter.

The referred change [2] is very recent so it is not fully
understood why the installation has been (?) working when
this project has the pyproject.toml. The reason is likely
that the similar case is wide spread among Python projects
and it has been sorted out somehow.

This is not fully understood how this goes. Good reading in
[1] and [3] related to latest updates on Python packaging on
top of PEP-517 and PEP-518.

[1] https://www.bernat.tech/growing-pain/
[2] https://github.com/pypa/setuptools/pull/1652
[3] https://www.bernat.tech/pep-517-518/

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>