Change REST API endpoints (breaking)
There was a bad design for REST API. There was colliding API
paths for {category}  and {keywords} which made it impossible
or error prone to decide which API endpoint was being used.

There was also bad design which added own API endpoints for
digest, tags, groups and uuid attributes. These search
attributes must be used through the content category API
endpoints.

The design is that the REST API manages collection of
resources through one API endpoint. The API endpoint can
be any of the content categories `snippet`, `solution` or
`reference`. All search queries are send through specific
API endpoint with query parameters.

This is not yet reflected to implementation.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>