Refactor resource attribute order
Because resources are operated in dictionaries, this is not
a breaking change. Default content in YAML files are stored
with attributes sorted.

The order of resource attributes were changed to be more
logical.

The `brief`, `description` and `name` describe the resource
and they are put close to each others. Same goes with the
`digest` and `uuid` which are now last in the list. The plan
is to count the digest over the whole resource eventually so
it is logical that the `digest` attribute is last.

The resource `category` is logically the first item which
defines the content. The `data` is attribute is the most
important part of the content so it is the second after
the `category` attribute followed by the descriptive
attributes brief, description and name.

The filename is considered last just before timestamps
for create and update times.

The `source` attribute is a link to source so it will be
still moved after `links` attribute.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>