Change resource attribute order in text templates
The same change with the resource attribute order is now
done also for text templates.

For example the descriptive attributes: brief, description
and name are grouped. Also the source attribute is now after
links in snippet text template. With the reference text
template, the source attribute is left in the same place as
with the snippet template. In case of reference template,
the first listed attribute is the links attribute. It was
considered that the descriptive attributes are more important
than the source attribute and thus the source was left in
the same place as with the snippet text template.

With the Markdown template, affected attributes are part
of the metadata. In case of Markdown metadata, attributes
are in sorted order and this change did not affect the
Markdown templates.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>