Fix Pylint warnings
Python third party modules for development were updated for
Sphinx next major version 2.0.0. The version 2 from Sphinx
does not support Python 2 anymore.

Also the Pyflakes was updated to next major version 2.1.1
from 1.6.0. This allowed update for Flake8 with pipenv that
had previously complained incompatible versions with Flake8
update. This had been stalling the Flake8 update previously.

This change revealed Pylint warnings with Python 2 which were
fixed or ignored. The external gzip generated warning that was
ignored. The indentation of few regexp matches to a visual
indentation but not to proposal generated in Python 2 and thus
this warning was disabled. That is, one code produces two
warnings in Python 2 Pylint which are conflicting. The one
that pleased authors eye was chosen and another was ignored.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>