Change REST API endpoints (breaking)
It is now possible to use generic {id} in the REST API
instead of {digest} or {uuid}. The internal logic for
the REST API server can now handle any given ID in the
API URL to find the resource.

This is part of the REST API refactoring that removes
colliding API endpoints. The target is to remove /uuid
and /digest endpoints and access all the resources from
content category API endpoints: /snippets, /resources
and /references.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>