Remove /digest, /uuid and /{sall} API endpoints
The digest and UUID specific API end points are not needed.
All access to content categories based on identity like
message digest or UUID can be done through /snippets,
/solutions or /references API endpoints.

The /{sall} was colliding API endpoint that was resulted
from a bad design.

The /tags and /groups API endpoints were left as is for
now. These will be changed. These API endpoints are meant
to manage ``tags`` and ``groups`` resources and not to
query resources from snippets, resources or groups API
endpoints. These two API endpoints will eventually support
only GET request which respond for example with unique
groups and tags.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>