Change HTTP resource response links and id
When HTTP requests is made for explicit resource with a
resource ID, the response links self and data id attributes
always use digest. This identity is considered better and
the main ID at the moment.

Previously the HTTP response used UUID with the response if
the request contained ``uuid`` query parameter. The problem
is that the REST API supports partial identities and it is
not possible to chose either digest or UUID based on HTTP
request at the moment. The reason is that in case the content
is updated and digest is used, the URI must be changed.

This change simplifies the logic by always using digest.

How ever, there is a feeling that changing the digest and
thus the URI when a resource is updated, is a bad practise.
There is no full understanding why this is bad at the moment.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>