Change server base path configuration option
The `--server-app-base-path` was renamed as `--server-base-path`.

The original thought was that Snippy would integrate security
services like authentication and authorization. This does not
make sense. If the tool is run in a cluster, there already is
some kind of an API GW that offers these servers.

This area is also complicated and there is no reason to start
to manage these services in Snippy tool. The secured HTTPS is
enough to provide a microservice for content management.

Because of this, the original plan to integrate auth API path
was removed. This allowed to remove the `app` from the server
base path.

If user management is wanted for some reason, it can be done
behind `/users` API endpoint in the same server base path than
rest of the API endpoints.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>