Fix searching content from PostgreSQL database
The commit that introduced the `identity` field to search
from digest or uuid attributes was not compatible with the
PostgreSQL. The SQL syntax had uuid LIKE <string> which
was constantly failing with the PostgreSQL.

It is likely that the Sqlite converts the UUID automatically
to string. In case of PostgreSQL, there is a UUID type with
real logic. Comparing string to UUID does not work in case
of PostgreSQL without explicitly casting the UUID to string
[1].

[1] https://stackoverflow.com/a/42449905

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>