Change version key-value mathematical operators
The operator `==` must be used instead of `=`. The later
operator `=` causes an error. This is in line with the
Mathematica operators [1].

On top of the Mathematica operators, the `~` operator
is supported. This meaning is left to client and it
is intended to be used in situations like "close to
version" or "not entirely sure but something like".

[1] https://en.wikipedia.org/wiki/Relational_operator

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>