Change REST API queries with UUID
The UUID must be now always used in full length in REST
API search queries. This is part of refactoring REST API
resource identifiers to be immutable [1].

Initially the design approach was to use mutable digest
that is counted based on resource attributes. If these
were changed, the digest and thus the URI was changed.
This not not a good practise [1].

This is part of larger refactoring to move towards
resource UUDI as URI ID. The resource external UUID is
immutable. There is no current plans to remove support
of digest based queries. But the REST API HTTP responses
that currently use digest in HTTP response id and links
attributes are going to change to use UUID.

In case of CLI based queries with UUID, the previous
implementation already forced full length match for
search UUID.

[1] https://stackoverflow.com/a/35845343

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>