Edit Python module versions for release
PostgreSQL database adapter psycopg2 version was updated.
This module is used with Snippy server as an option to
psycopg2cffi for PostgreSQL adapter.

The main use case is to run the Snippy server in container
which uses Alpine Linux which does not yet support the
latest version from psycopg2. This was the reason why the
Dockerfile version is still having version requirements
psycopg2>=2.7.5,<=2.8.1.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>