Fix install requirement for PyYAML version
Previously the PyPI installer required PyYAML version between
few releases. There are how ever multiple new versions from
PyYAML and recent Linux release may have anything between 3.*,
4.* or 5.*. This was a problem if the Snippy was installed
as a system package. The recommended way to install is to
install the package for local user, but even that may pose
problems due to various versions.

The user group at the moment is the CLI users. The explicit
PyYAML made it hard to install.

The tool uses the secure version from the load (safe_load
and safe_dump) so the security aspect is taken care with
any version from the PyYAML.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>