Change REST API JSON schema validation
HTTP requests are now validated against JSON schema that
is auto generated from Swagger/OAS specification.

This removes the support for resource attributes that had
previously supported array or string formats. For example
the links attribute could be sent in array or string format.
The reason to remove this is to simplify the (and thus
reduce errors) the JSON schema definition.

Following notes and updates are going to be done later:

  - Setting resource attributes to null is broken in this
    commit. This would require 'nullable' property that is
    supported only in OAS 3.

  - Additional resource attributes like uudi in HTTP
    requests does not yet generate error. This is going
    to be changed.

  - There will be command line option to disable JSON
    schema validation for HTTP requests.

  - For internal testing, there will be command line
    option to enable JSON schema validation for sent
    HTTP methods. This tries to make sure that clients
    can work with the provided JSON schema.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>