Change accepted resource attributes in HTTP request
All HTTP request resource attributes are now validated so
that additional properties are not allowed. This change
requires updating test case default data management since
the default data always contain all the attributes which
is now causing error.

There is ongoing refactoring also to improve readability
of tests with the same change.

This is a partial commit and tests will fail.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>