Fix test error validation Python 3.4
The Python 3.4 seems to have different length of JSON
schema validation error than with other Python versions.

Root cause was not solved. The jsonschema is pinned so
it is the same in different Python versions. It is likely
something to do with special characters like Unicode between
Python 2 and 3 or other dependency in the chain that causes
different result.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>