Fix Dockefile layout and documentation
The Dockerfile was refactored and intensive documentation
with separate TODO list was added. Also .dockerignore was
added to prevent adding unnecessary files to Docker images.

The reason to document the Dockerfile was to clarify the
security aspects of Dockerfiles. Also the documentation
tries to allow evaluating the implementation of Dockerfile
from security point of view.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>