Fix Docker image run example with host network
There is now example how to run the Docker image with host
network. This is not recommended configuration even though
it is supported because it is a convenient way to run the
tool locally.

The container image healt check was fixed so that it should
now support HTTP or HTTPS protocols.

The separate configuration for the server IP and port were
removed because those are not needed anymore. One reason
is the curl supports hostname based requests instead of
space separated IP and port like with nc tool that was
previously used.

Also the configuration possibility was removed from the
EXPOSED port because it was not really needed. See further
explanation in the documentation part.

Installing curl increased the image size only slightly
with 0.5MB.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>