Change REST API default base path
Renamed the default REST API base path. The previous base
path contained 'app' which was from old design that used
also 'auth'. The base path is now shorter and it contains
'rest'. The thought is that there may be 'graphql' someday.

Also the version was removed from the base path. There
seems to be only wrongs ways to version API paths [1].
Without version in URI is preferred but it brings problems
when versioning is needed. Bigger applications tend to use
the version but it brings extra maintenance (docs, code,
test) to support all the versions. Snippy is a small
application and it may be that there will be Graphql API
before there will be v2 from the REST API.

If versioning is needed, it is possible to use subdomains
like http://api1.snippy.rest and http://api2.snippy/rest
which lead to different version of Snippy server. This is
much easier for small applications like this.

old: /snippy/api/app/v1
new: /api/snippy/rest

[1] https://www.troyhunt.com/your-api-versioning-is-wrong-which-is/

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>