Remove REST API route /snippy
All REST API routes that the tool handle must start with
the REST API base path. There may be some conflicts if
for example responses would be send from the API root.

Like http://127.0.0.1:8080. If there is a single instance
from API GW like nginx, it should know to which backend
host to forward the request. This may cause incorrect
requests sent to Snippy server so HTTP 404 must be a
response in all cases where the API path does not start
with the base path.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>