Fix Gunicorn server bind address syntax
The bind address was set in list context and it used complex
syntax that was a left over from times when IP address and
host were separate parameters.

Gunicorn supports string and list context for bind address.
But in case if Snippy, there is only one address so it is
much clearer just to give the host as a string.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>