Fix server bind IP address in Docker container
The server used to always bind to 0.0.0.0 in Docker container.
This is a security risk. Now the container tries to read the
container runtime hostname and related IP address.

There is a fallback to address 0.0.0.0 if reading of the host
fails. There is a security level log message if this was the
outcome from container IP address reading.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>