Fix server startup in Docker container
If all server configuration options were defined from
environment variables and no command line options were
given for the server in container, the server startup
failed. The code logic did not look to environment
variables to determined if the server is going to run

Now it is possible to have all configuration options
in environment variables and start the server.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>