Fix Docker container health check
The healthcheck was broken when special tag 'container.
hostname' was added. The previous logic was that the
Snippy server would read this and update, but that did
not work if defaults were use. That is, the server was
going to bind to default container hostname IP. In this
case the IP was extracted by the code but it was not
available for the Docker healthcheck.

The root of the problem is:

1. The target user group is a casual user with only
   Docker to run containers.

2. Binding to 0.0.0.0 is considered security risk
   that it is not wanted to do.

This forces the implementation to seek the container
internal IP based on hostname. It is considered that
it is too difficult to give this configuration setting
to user that starts the container. The intention is
that this just needs to work without too much command
line options and hacks from 'docker run'

See more information details from the update Dockerfile
documentation part.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>