Fix log message format when JSON logs enabled
There were two logs related to server startup and stop that
were printed as plain strings even when the JSON log format
was enabled. This was a possible problem for server. It may
be that the logs are post processed as JSON (when set) and
plain text logs may break, cause error or log may be just
discarded when post processing of logs is made.

All logs are now in JSON format when --log-json is enabled.
This is enabled by default in server based on container.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>