Fix running CLI commands with container
Previous changes broke the possibility to run command line
options with the Docker container. This is now fixed. There
is a supported use case to use the Docker container as a
CLI interface. See the Dockerfile documentation part for
more information.

Fix failure to read bool environment variable. Values that
should have evaluated to False were evaluated internally
to True. Environment variables are strings and string '0'
will evaluate to True instead of False. For boolean values
only supported values for True '1' and 'True'. For false
supported values are '0' and 'False'. The code has bit
wider support for few other alternatives but to keep things
simple, these are the values that must be used.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>