Test the server in Docker container
There is now a test infrastructure that allows testing use
cases for Docker container based server. There is only one
initial test but this should allow repeating the same Pytest
fixtures.

The fixtures allow exception handling the fixture and there
is no need to handle them in the test. Using fixture also
hides the actual implementation (Pythons Docker SDK) behind
the fixture.

The tests are actually pretty fast and it is possible to
run Docker tests in parallel. But since these tests require
working Docker and run privileges, the tests are not run
by default and there are new make targets to run the tests.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>