Fix Makefile install target for PEP517
The latest pip and setuptools prevent --editable installation
for pyproject.toml based builds. When the latest setuptools
and pip were installed, various make install targets failed.
The --editable flag and --user flags are removed.

The make install targets are meant to run inside a virtual
environment. The --user flag does not make sense in venv and
thus it was removed. In case of virtual environment, the
installation always goes to the virtual environment. Note
that there is no check in the Makefile if virtual environment
is activate. This means that if the commands are run without
it, the install goes to global if not run with underprivileged
user.

Fix Makefile install target names. All the install targets now
have 'install' prefix.

Fix the $(PIP) because it is better to use '$(PYTHON) -m pip'.
This always guarantees the correct pip related to used Python.

Fix the Makefile. Move clean target at the end and remove
phony targets that are not needed.

Add target to upgrade Python setuptools, wheels and twine.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>