Add own Docker healthcheck for Snippy server
There is now '--server-healthcheck' option that does a
healthcheck against Snippy server. This allows parsing
the Docker container server bind IP address exactly the
same in the server and in the healtcheck. This method
avoids start-up scripts and parsing the Docker container
bind IP address to a separate file.

The embedded healthcheck also seems to survive through
performance test where the server is loaded. The curl
based method moved the container to faulty state even
though the performance test was successful.

Shortly, the problems are below and the only proper
solution to do this is to have own service (separate
or embedded) to do the healthcheck. Embedded service
as implemented here is easier to manage.

The problem is that the sever bind IP is:

  1) Not know in Docker image compile time.
  2) Not easily available for user when Docker container is started.
  3) Not secured when server binds to 0.0.0.0.
  4) Not available easily inside container from hostname.
  5) Not easily available for curl/nc based healthcheck.

See more information from the Dockerfile documentation
section.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>